Consolidate supabase client module
