+++ src/utils/cloud.js
@@ -1,5 +1,5 @@
 // src/utils/cloud.js
-import { supabase } from "../supabaseClient";
+import { supabase } from "./supabaseClient";
 
 /* ------- helpers ------- */
 const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
