+++ src/utils/supabaseClient.js
@@ -1,4 +1,3 @@
-// src/supabaseClient.js
 import { createClient } from "@supabase/supabase-js";
 
 const url = import.meta.env.VITE_SUPABASE_URL;
